fix: replace current history on search
Currently, each keystroke being made on search would result in a new history entry being pushed, which is unideal, because it means that the back/forward entries are filled up with unnecessary entries. This also affects the app's own back button that is present in some screens, the button would just be undoing what you've typed in search earlier instead of actually going to the previous screen, which is more ideal.

This change makes it so that the current history entry is being replaced instead, using `replaceIn` instead of the default `pushIn` update type.