Properly sort chapter list by chapter number
The sort order was switched and instead of ASC it was sorted DESC and vice versa.
This was caused because the server returned the list (most of the time) in ASC order by chapter number.

Additionally, the  list was overall never properly sorted by chapter number and only worked because the server (most of the time) returned it in the correct order by chapter number.
However, this is not guaranteed at all since the server sorts in ASC order by the chapter id by default