Change order of update manga migration
This prevents issues of manga having a specific library state while they are incorrectly in or not in categories.

E.g. a manga was removed from the library but the category update mutation failed, which now causes it to still be in the set categories while not being in the library anymore