Fix previous preview version mappings
Update ui versions to the latest compatible versions, they were pointing to the first version that updated the minimum server version while they needed to point to the previous commit of the commit that changed the minimum server version to a higher version.

Updated ui versions that pointed to non-existing revisions.
Multiple commits at once were pushed to the master which caused the revision of the mapping update commit to not be a release