Normalize language code comparisons
Due to using the browsers' preferred languages by default, there was a possibility that the preferred language codes did not match any extension/source language, which lead to broken filtering.

While this happened, it was also possible that the preferred languages still matched properly to a language in the language filter, which caused confusion.
This was the case e.g., for "en-GB" or "en-US"