Prevent white screen on page load
Bug introduced with ff58dcc4803e23207234e7832f27bbb950a76c74.
With 18ddf46f17d7d88276349d21208bbaad076127ce the "appearance" settings got moved from the local storage to the server metadata.
However, the old settings still remained in the local storage, which includes a value with they key "appTheme". This was the id of the selected app theme.
With ff58dcc4803e23207234e7832f27bbb950a76c74 this key got resued for the app theme object, which leads to a type error in case the old setting is still present in the local storage.