Fix handling of disabled manga migration steps
In case e.g., the "tracking" should not get migrated, the "migration action" was still created.
While creating the action, an error was thrown because the "tracking" information was missing, as expected, since it's not required.

To prevent this, the actions should only get created in case they are actually required for the migration.