Optionally require forced page refresh on detected webui update
A forced refresh is only necessary in case the webui was already loaded and thus, still uses outdated files from the old version.

In case the app just got freshly opened in a tab, a refresh is not necessary because the latest files are already being used.
In that case, just show an update dialog which won't force a tab refresh.